requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
idna==3.10
imageio==2.37.0
joblib==1.4.2
kagglehub==0.3.9
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.0
networkx==3.4.2
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
threadpoolctl==3.5.0
tifffile==2025.2.18
tqdm==4.67.1
tzdata==2025.1
urllib3==2.3.0
